Add files via upload
these files contains the holds the content to open bmp image files. This current version supports a bmp image of  MAX size 40x24 (width x height). Lesser size is also supported, but for proper interpretation of the bmp image, it must be a multiple 4 for both the width and height of the image. This BMP image is 4 bit multi-colour (16 colours) supported. The image generated here is a raster graphics NOT actual pixel drawing to the screen.